+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## master
 
+- **Important** The `ClassOf` alias for `createClass` has been removed, us `createClass` (with new generics) directly
+
 Changes:
 
 - `createType<T = Codec>('...': K)` now returns `Codec` when `K` is not in the InterfaceRegistry, with `T` override